libc: use 64-bit clock types, fix usleep behavior, fix Malloc.unit_aligned for 64-bit arch

+++ userspace/libraries/libc/src/include/stddef.h
@@ -21,5 +21,5 @@ typedef unsigned char wchar_t;
 
 typedef unsigned long long time_t;
 typedef unsigned long long suseconds_t;
-typedef unsigned long useconds_t;
-typedef unsigned long clock_t;
+typedef long long useconds_t;
+typedef unsigned long long clock_t;
